add webpack config
