+++ package.json
@@ -1,0 +1,42 @@
+{
+  "name": "js-tic-tac-toe",
+  "version": "1.0.0",
+  "description": "A collaborative project of a tic-tac-toe game aimed at allowing two users to play against each other. The game is capable of ending in 3 states: with a winner, loser, and a draw. Each user after entering their names are given a symbol that identifies them on the board.",
+  "main": "index.js",
+  "dependencies": {
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb-base": "^14.1.0",
+    "eslint-plugin-import": "^2.20.2",
+    "nes.css": "^2.3.0",
+    "stylelint": "^13.3.3",
+    "stylelint-config-standard": "^20.0.0",
+    "stylelint-csstree-validator": "^1.8.0",
+    "stylelint-scss": "^3.17.2"
+  },
+  "devDependencies": {
+    "css-loader": "^5.0.0",
+    "css-minimizer-webpack-plugin": "^1.1.5",
+    "file-loader": "^6.2.0",
+    "mini-css-extract-plugin": "^1.2.1",
+    "webpack": "^5.3.2",
+    "webpack-cli": "^4.1.0"
+  },
+  "scripts": {
+    "test": "jest",
+    "build": "webpack --mode development",
+    "build:prod": "webpack --mode production",
+    "start": "webpack-dev-server --host 0.0.0.0 --port 7000",
+    "watch": "webpack --watch"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ferbaco86/js-tic-tac-toe.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/ferbaco86/js-tic-tac-toe/issues"
+  },
+  "homepage": "https://github.com/ferbaco86/js-tic-tac-toe#readme"
+}
